% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_returns}
\alias{plot_returns}
\title{Plot returns}
\usage{
plot_returns(rets)
}
\arguments{
\item{rets}{periodic returns.}
}
\value{

}
\description{
Plot returns
}
\examples{
x <- get_prices(c("AAPL", "IBM"), from = "2010-01-01")
rets <- daily_returns(x)
plot_returns(rets)

}
