% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_weights}
\alias{plot_weights}
\title{Plot weights}
\usage{
plot_weights(h_weights, limits = c(0, 1))
}
\arguments{
\item{h_weights}{xts object of historical portofio weights.}
}
\value{

}
\description{
Plot weights
}
