% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_class.R
\name{portfolio}
\alias{portfolio}
\title{Portfolio Constructor}
\usage{
portfolio(symbols = NULL, prices = NULL, weights = NULL, ...)
}
\arguments{
\item{x}{an xts object.}
}
\value{
a porftolio object
}
\description{
Portfolio Constructor
}
